splitting the analyzers to respective features
